all new clientprotocol methods work through namenode protobuf now
